dev-ruby/responds_to_parent: add version for ruby22 and rails42
Cobble together a version that works with Rails 4.2.x and Ruby 2.2.
The previous ebuild version does work with Rails 4.2.x but not with
Ruby 2.2 since it tries to modify NilClass which is now frozen in Ruby 2.2.

This version is based on the latest upstream commit + a Rails 4.2 patch
from a fork mentioned in the github network.

Package-Manager: portage-2.3.0